Add fuzzymatching to, and remove anonymity from, the hug command

+++ services/hug/utils.py
@@ -1,29 +1,61 @@
 import discord
 import re
+from services.utils import check_perms
+from fuzzywuzzy import fuzz
 
 
 # TODO Implement logging if/when a log handler is added
 async def remove_invocation(cog, ctx):
-    try:
-        await ctx.message.delete()
-    except discord.Forbidden:
-        print("Invalid permissions: Bot needs manage_messages to delete user messages.")
-    except discord.NotFound:
-        print("Message not found. Something went wrong during command invocation.")
-    except discord.HTTPException as e:
-        print(f"HTTP Exception: {e}")
+    if check_perms(ctx, discord.Permissions(
+        manage_messages = True
+    )):
+        try:
+            await ctx.message.delete()
+        except discord.Forbidden:
+            print("Invalid permissions: Bot needs manage_messages to delete user messages.")
+        except discord.NotFound:
+            print("Message not found. Something went wrong during command invocation.")
+        except discord.HTTPException as e:
+            print(f"HTTP Exception: {e}")
 
 
 async def find_user(guild, user):
-    match = re.search(r'\d+', user)
+    is_mention = re.search(r'<@!\d+>', user)
 
-    if not match:
-        raise InvalidMentionException()
-
-    member = guild.get_member((int(match[0])))
+    if is_mention:
+        user_id = await convert_mention_to_id(user)
+        member = guild.get_member(user_id)
+    else:
+        member = await fuzzymatch_name(guild, user)
 
     return member
 
 
-class InvalidMentionException(Exception):
+async def convert_mention_to_id(user):
+    match = re.search(r'\d+', user)
+
+    if not match:
+        raise InvalidUserException
+
+    return int(match[0])
+
+
+async def fuzzymatch_name(guild, user):
+    """
+        If the supplied user string is at least 60% similar to an existing guild user, consider that a match
+    """
+    best_match = (None, -1)  # Member object and fuzzy ratio
+    for member in guild.members:
+        if member.nick is not None:
+            ratio = fuzz.partial_ratio(user.lower(), member.nick.lower())
+        else:
+            ratio = fuzz.partial_ratio(user.lower(), member.name.lower())
+
+        if ratio > best_match[1]:
+            best_match = (member, ratio)
+
+    return best_match[0]
+
+
+class InvalidUserException(Exception):
     pass
